backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
python-multipart==0.0.9
opencv-python
facenet-pytorch
torch
torchvision
scipy
pillow
qrcode
pyzbar
